IPRT/rest: build fix. bugref:9167
svn:sync-xref-src-repo-rev: r124971

+++ src/VBox/Runtime/testcase/tstRTCRest-1.cpp
@@ -1,4 +1,4 @@
-/* $Id: tstRTCRest-1.cpp 74195 2018-09-11 11:45:35Z knut.osmundsen@oracle.com $ */
+/* $Id: tstRTCRest-1.cpp 74196 2018-09-11 11:48:16Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT Testcase - REST C++ classes.
  */
@@ -1718,7 +1718,7 @@ static void testArray()
                 else
                     RTTESTI_CHECK_RC(Arr4.prepend(pDirect), VINF_SUCCESS);
             }
-            RTTESTI_CHECK(Arr4.size() == i + 1);
+            RTTESTI_CHECK((ssize_t)Arr4.size() == i + 1);
             RTTESTI_CHECK(Arr4.isEmpty() == false);
         }
